% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_yasso.R
\name{run_yasso}
\alias{run_yasso}
\title{Call the Fortran version of YASSO15}
\usage{
run_yasso(par, sdl)
}
\arguments{
\item{par}{model parameters}

\item{sdl}{input data in a list}
}
\value{
Simulated soil carbon
}
\description{
Call the Fortran version of YASSO15
}
\examples{
soilC <- run_yasso(par = sample_parameters, sdl = sample_data_list)
}
