% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_cal.R
\docType{data}
\name{sample_data_cal}
\alias{sample_data_cal}
\title{Sample input data}
\format{A \code{tibble} with 2 cols and stuff:
\describe{
\item{x}{random normal data}
\item{y}{data that is normal and correlated with \code{x}}
}}
\usage{
sample_data_cal
}
\description{
A list withsample data.
}
\keyword{datasets}
