% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_parameters.R
\docType{data}
\name{sample_parameters}
\alias{sample_parameters}
\title{Sample parameters}
\format{
A named numeric vector with the following contents:
\describe{
\item{aA, aW, aE, aN}{Decomposition rates of AWEN pools}
\item{pXY}{Relative mass flow from pool X to Y}
\item{w1, w2, w3, w4, w5}{Leaching correction for size classes}
\item{b1, b2}{Temperature dependence of the AWE pools}
\item{bN1, bN2}{Temperature dependence of the N pool}
\item{bH1, bH2}{Temperature dependence of the H pool}
\item{g}{Precipitation dependence of the AWE pools}
\item{gN}{Precipitation dependence of the N pool}
\item{gH}{Precipitation dependence of the H pool}
\item{aH}{Relative mass flow to the H pool}
\item{pH}{Decomposition rate of the H pool}
\item{th1, th2, r}{Woody size dependence}

}
}
\usage{
sample_parameters
}
\description{
A sample parameter set for the YASSO model.
}
\keyword{datasets}
