% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_yasso.R
\name{calibrate_yasso}
\alias{calibrate_yasso}
\title{Call the Fortran version of YASSO15}
\usage{
calibrate_yasso(par, sdl)
}
\arguments{
\item{par}{model parameters}

\item{sdl}{input data in a list}
}
\value{
Simulated soil carbon
}
\description{
Call the Fortran version of YASSO15
}
\examples{
soilC <- calibrate_yasso(par = sample_parameters, sdl = sample_data_cal)
}
