% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_yasso.R
\name{run_yasso}
\alias{run_yasso}
\title{Run the YASSO model}
\usage{
run_yasso(par, sdl)
}
\arguments{
\item{par}{A numeric vector of YASSO parameters.}

\item{sdl}{A list of input data. See \code{\link{sample_data_run}} for
details.}
}
\value{
A matrix containing the initial soil carbon on the first row and
  simulated soil carbon on the following rows.
}
\description{
\code{run_yasso()} runs the YASSO model and returns simulated soil carbon.
}
\details{
\code{run_yasso()} wraps the Fortran90-release of the soil carbon model
YASSO15 into a simple R-function. The function is a convenient way to call
the Fortran-release, which is two orders of magnitude faster than
the R-release.
}
\examples{
soilC <- run_yasso(par = sample_parameters, sdl = sample_data_run)
}
