% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_yasso.R
\name{run_yasso}
\alias{run_yasso}
\title{Run the YASSO model}
\usage{
run_yasso(
  par,
  n_runs,
  time,
  temp,
  prec,
  init,
  litter,
  wsize,
  leac,
  sspred = 0L
)
}
\arguments{
\item{par}{A numeric vector of YASSO parameters.}

\item{n_runs}{Input data. Refer to \code{\link{sample_data_run}} for now.}

\item{time}{-||-}

\item{temp}{-||-}

\item{prec}{-||-}

\item{init}{-||-}

\item{litter}{-||-}

\item{wsize}{-||-}

\item{leac}{-||-}

\item{sspred}{Optional integer, should steady state mode be used (1 = yes).}
}
\value{
A matrix containing the initial soil carbon on the first row and
  simulated soil carbon on the following rows.
}
\description{
\code{run_yasso()} runs the YASSO model and returns simulated soil carbon.
}
\details{
\code{run_yasso()} wraps the Fortran90-release of the soil carbon model
YASSO15 into a simple R-function. The function is a convenient way to call
the Fortran-release.

The function provides YASSO with the initial soil carbon values in the vector
\code{init} and runs the model one time step at a time. The simulated
carbon of the current time step is used as the initial value of the next time
step. The model runs until it has looped over all the time steps.
}
\examples{
soil_c <- run_yasso(
 par = sample_parameters,
 n_runs = sample_data_run$n_runs,
 time = sample_data_run$time,
 temp = sample_data_run_Tmonthly$temp,
 prec = sample_data_run_Tmonthly$prec,
 init = sample_data_run$init,
 litter = sample_data_run$litter,
 wsize = sample_data_run$wsize,
 leac = sample_data_run$leac
)
}
