% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_yasso.R
\name{run_yasso}
\alias{run_yasso}
\title{Run the YASSO model}
\usage{
run_yasso(
  par,
  n_runs,
  time,
  temp,
  prec,
  init,
  litter,
  wsize,
  leac,
  sspred = 0L
)
}
\arguments{
\item{par}{\code{double} A vector containing the YASSO parameters.}

\item{n_runs}{\code{integer} Number of time steps to run the model over.}

\item{time}{\code{double} Length of each time step in years.}

\item{temp}{\code{matrix} Average temperature for each month of each modelled
year.}

\item{prec}{\code{double} Annual precipitation sum of each year.}

\item{init}{\code{double} Initial soil carbon values (A, W, E, N, H) to start
the simulation with.}

\item{litter}{\code{matrix} Litter input to the model on each time step (A,
W, E, N, H).}

\item{wsize}{\code{double} Litter size on each time step.}

\item{leac}{\code{double} Leaching correction factor on each time step. Only
defined for litter bag measurements - describes how much litter falls out of the holes in the bags over time. Usually a calibrated value.}

\item{sspred}{\code{integer} Defines if steady state mode should be used (1 =
yes).}
}
\value{
A matrix containing the initial soil carbon on the first row and
  simulated soil carbon on the following rows.
}
\description{
Run the YASSO model and return simulated soil carbon.
}
\details{
The \code{run_yasso()} function is designed for generic use, such as making
predictions with YASSO15. The user provides YASSO15 with driver data and
initial carbon in a vector. The model “rolls” the carbon forward one time
step at a time using the simulated carbon of the current time step as the
initial carbon of the next step. The model runs until it has looped over all
the time steps. See also the sample dataset \code{\link{sample_data_run}}.
}
\examples{
soil_c <- run_yasso(
 par = sample_parameters,
 n_runs = sample_data_run$n_runs,
 time = sample_data_run$time,
 temp = sample_data_run$temp,
 prec = sample_data_run$prec,
 init = sample_data_run$init,
 litter = sample_data_run$litter,
 wsize = sample_data_run$wsize,
 leac = sample_data_run$leac
)
}
\seealso{
\code{\link{sample_data_run}}
}
